backend/requirements.txt
Django>=6.0
djangorestframework>=3.17
django-cors-headers>=4.0
mysqlclient>=2.2
